Fixed size types based on snej's pull request. https://github.com/snej/msgpack-c/commit/1784e7a3f3d012e1149adc9b8a3abb3646eb616f

+++ example/protocol.cc
@@ -17,7 +17,7 @@ namespace myprotocol {
 
 	struct Put : define< tuple<uint32_t, std::string, raw_ref> > {
 		Put() { }
-		Put(uint32_t f, const std::string& k, const char* valref, size_t vallen) :
+		Put(uint32_t f, const std::string& k, const char* valref, uint32_t vallen) :
 			define_type(msgpack_type( f, k, raw_ref(valref,vallen) )) { }
 		uint32_t&    flags() { return get<0>(); }
 		std::string& key()   { return get<1>(); }
